Add network block device boot support
This changeset implements boot with a network block device.
Initial steps are equal to tftp+nfs but instead of nfs this
method depends on an nbd-client started on the target within
the initrd. The rootfs is served on the slave through xnbd-server.

This setup is needed when the target system needs e.g.
extended file attributes like security labels to boot.
In this case cpio.gz and nfs do not work and both the initrd and
the rootfs need to be a full filesystem (e.g. ext4).

Boot sequence:
 DL -> tftp -> initrd (nbd-client+pivot_root) -> rootfs(nbd)

The implementation reuses as much as possible from the
existing tftp+nfs workflow. The startup of the xnbd-server
and its proper teardown required the use of an action/protocol
combination (similar to lxc).

The xnbd-server will use a random port (unless specified).
A helper function has been added to network.py.

Change-Id: Ibbaa10b4bdab49d9cc95d9b1c6cb3d67f848760a
Signed-off-by: Jan-Simon Möller <jsmoeller@linuxfoundation.org>